# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * منصة مقايضة - Finance Swap Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type FinancingRequestStatus =
  (typeof FinancingRequestStatus)[keyof typeof FinancingRequestStatus];

export const FinancingRequestStatus = {
  pending: "pending",
  active: "active",
  approved: "approved",
  closed: "closed",
} as const;

export interface FinancingRequest {
  id: number;
  fullName: string;
  phone: string;
  employer: string;
  salary: number;
  currentDebt: number;
  remainingMonths: number;
  bankName: string;
  sector?: string;
  financingPurpose?: string;
  financingType?: string;
  preferredFeature?: string;
  notes?: string;
  status: FinancingRequestStatus;
  offersCount: number;
  createdAt: string;
}

export interface CreateRequestInput {
  fullName: string;
  phone?: string;
  employer: string;
  sector?: string;
  financingPurpose?: string;
  financingType?: string;
  salary: number;
  currentDebt: number;
  remainingMonths: number;
  bankName?: string;
  notes?: string;
  preferredFeature?: string;
}

export type OfferStatus = (typeof OfferStatus)[keyof typeof OfferStatus];

export const OfferStatus = {
  pending: "pending",
  client_accepted: "client_accepted",
  accepted: "accepted",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Offer {
  id: number;
  requestId: number;
  advisorId: number;
  advisorName: string;
  advisorCompany: string;
  profitRate: number;
  monthlyInstallment: number;
  totalAmount: number;
  durationMonths: number;
  notes?: string;
  status: OfferStatus;
  createdAt: string;
}

export interface CreateOfferInput {
  advisorId: number;
  profitRate: number;
  monthlyInstallment: number;
  totalAmount: number;
  durationMonths: number;
  notes?: string;
}

export interface Advisor {
  id: number;
  name: string;
  company: string;
  phone: string;
  email?: string;
  licenseNumber: string;
  offersCount: number;
  successRate: number;
  createdAt: string;
}

export interface CreateAdvisorInput {
  name: string;
  company: string;
  phone: string;
  email?: string;
  licenseNumber: string;
}

export interface PlatformStats {
  totalRequests: number;
  totalOffers: number;
  totalAdvisors: number;
  approvedDeals: number;
  pendingRequests: number;
}

export type ListRequestsParams = {
  status?: ListRequestsStatus;
};

export type ListRequestsStatus =
  (typeof ListRequestsStatus)[keyof typeof ListRequestsStatus];

export const ListRequestsStatus = {
  pending: "pending",
  active: "active",
  approved: "approved",
  closed: "closed",
} as const;