# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: منصة مقايضة - Finance Swap Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offers/{id}/accept:
    post:
      operationId: acceptOffer
      tags: [offers]
      summary: Accept an offer (user action)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Offer accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/{id}/approve:
    post:
      operationId: approveOffer
      tags: [admin]
      summary: Admin approves a deal after user accepts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deal approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
...
components:
  schemas:
...
    Offer:
      type: object
      properties:
        id:
          type: integer
        requestId:
          type: integer
        advisorId:
          type: integer
        advisorName:
          type: string
        advisorCompany:
          type: string
        profitRate:
          type: number
        monthlyInstallment:
          type: number
        totalAmount:
          type: number
        durationMonths:
          type: integer
        notes:
          type: string
        status:
          type: string
          enum: [pending, client_accepted, accepted, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - requestId
        - advisorId
        - advisorName
        - advisorCompany
        - profitRate
        - monthlyInstallment
        - totalAmount
        - durationMonths
        - status
        - createdAt
...